+++ deployments/liqo/files/liqo-controller-manager-ClusterRole.yaml
@@ -624,14 +624,6 @@ rules:
   - update
   - watch
 - apiGroups:
-  - offloading.liqo.io
-  resources:
-  - quotas
-  verbs:
-  - get
-  - list
-  - watch
-- apiGroups:
   - rbac.authorization.k8s.io
   resources:
   - clusterrolebindings
@@ -788,12 +780,4 @@ rules:
   - list
   - patch
   - update
-  - watch
-- apiGroups:
-  - virtualkubelet.liqo.io
-  resources:
-  - vkoptionstemplates
-  verbs:
-  - get
-  - list
   - watch
